5.02

+++ src/main/java/jp/co/sony/ppog/controller/CentreController.java
@@ -77,7 +77,7 @@ public class CentreController {
 	 */
 	@GetMapping(value = "/city/{id}")
 	@ResponseBody
-	public RestMsg getCityInfo(@PathVariable("id") final Integer id) {
+	public RestMsg getCityInfo(@PathVariable("id") final Long id) {
 		final CityInfoDto cityInfoDto = this.centreLogicService.getCityInfoById(id);
 		return RestMsg.success().add("citySelected", cityInfoDto);
 	}
@@ -90,7 +90,7 @@ public class CentreController {
 	 */
 	@GetMapping(value = "/nations/{id}")
 	@ResponseBody
-	public RestMsg getListOfNationsById(@PathVariable("id") final Integer id) {
+	public RestMsg getListOfNationsById(@PathVariable("id") final Long id) {
 		final List<String> nations = this.centreLogicService.getListOfNationsById(id);
 		return RestMsg.success().add("nationsWithName", nations);
 	}
@@ -129,7 +129,7 @@ public class CentreController {
 	 */
 	@DeleteMapping(value = "/city/{id}")
 	@ResponseBody
-	public RestMsg deleteCityInfo(@PathVariable("id") final Integer id) {
+	public RestMsg deleteCityInfo(@PathVariable("id") final Long id) {
 		this.centreLogicService.removeById(id);
 		return RestMsg.success();
 	}
